requirements.txt
apache-beam==2.29.0
asgiref==3.2.10
astroid==2.4.2
attrs==20.3.0
autopep8==1.5.5
avro-python3==1.9.2.1
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
crcmod==1.7
dill==0.3.1.1
Django==3.0.8
django-cors-headers==3.5.0
django-countries==7.1
django-extensions==3.1.1
django-filter==2.3.0
django-impersonate==1.7
django-reset-migrations==0.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
docutils==0.15.2
Faker==5.8.0
fastavro==1.4.0
flake8==3.9.1
greenlet==0.4.17
gunicorn==20.0.4
hdfs==2.6.0
idna==2.9
importlib-metadata==3.4.0
iniconfig==1.1.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
oauth2client==4.1.3
packaging==20.8
pep8==1.7.1
Pillow==8.2.0
pluggy==0.13.1
protobuf==3.16.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydot==1.4.2
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.5.3
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
validate-docbr==1.8.2
wrapt==1.12.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.1.2
